+++ src/index.ts
@@ -6,4 +6,6 @@ export { default as useFormContext } from './useFormContext';
 
 export { default as useField } from './useField';
 
+export { default as useCheckboxField } from './useCheckboxField';
+
 export { get } from './utils';