+++ CHANGELOG.md
@@ -1,10 +1,16 @@
+## Unreleased
+
+- Removed options from `useField` hook
+- Removed `Field` and `FieldOptions` types
+- Added `useCheckboxField` hook
+
 ## 0.2.1 (April 29, 2020)
 
-- Fix overriding onSubmit on Form component
+- Fixed overriding onSubmit on Form component
 
 ## 0.2.0 (April 29, 2020)
 
-- Move submit to form instance
+- Moved submit to form instance
 
 ## 0.1.0 (April 29, 2020)